update:컴포넌트 화
