+++ src/containers/app/App.js
@@ -9,20 +9,13 @@ import {
   Text
 } from 'react-native';
 
-const config = {
-  apiKey: "AIzaSyBBAvfR4ffQmKtg2GQTqAi3htH3R2ehvqw",
-  authDomain: "roompackersdb.firebaseapp.com",
-  databaseURL: "https://roompackersdb.firebaseio.com",
-  storageBucket: "roompackersdb.appspot.com",
-  messagingSenderId: "332389298611"
-}
 
 import LoginPage from '../loginPage/LoginPage'
 
 class App extends Component {
   render() {
     return (
-      <LoginPage config={config} />
+      <LoginPage/>
     );
   }
 }
